Remove unnecessary package name usage
