+++ library/runtime-api/src/commonMain/kotlin/io/matthewnelson/kmp/tor/runtime/api/builder/SocksFlagBuilder.kt
@@ -15,12 +15,12 @@
  **/
 @file:Suppress("FunctionName", "PropertyName")
 
-package io.matthewnelson.kmp.tor.runtime.api.config.builders
+package io.matthewnelson.kmp.tor.runtime.api.builder
 
 import io.matthewnelson.kmp.tor.core.api.annotation.KmpTorDsl
 import io.matthewnelson.kmp.tor.runtime.api.ThisBlock
 import io.matthewnelson.kmp.tor.runtime.api.apply
-import io.matthewnelson.kmp.tor.runtime.api.config.TorConfig
+import io.matthewnelson.kmp.tor.runtime.api.TorConfig
 import kotlin.jvm.JvmField
 import kotlin.jvm.JvmSynthetic
