+++ library/runtime-api/src/commonTest/kotlin/io/matthewnelson/kmp/tor/runtime/api/ControlPortUnitTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Matthew Nelson
+ * Copyright (c) 2024 Matthew Nelson
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.runtime.api.config
+package io.matthewnelson.kmp.tor.runtime.api
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class HSNumIntroductionPointsUnitTest {
+class ControlPortUnitTest {
 
     @Test
-    fun givenLambdaClosure_whenConfigurationAttempt_thenDoesNothing() {
-        val map = mutableMapOf("3" to 3)
-        var builder: TorConfig.HiddenServiceNumIntroductionPoints? = null
-        TorConfig.HiddenServiceNumIntroductionPoints.configure(map) {
-            builder = HSv3(points = 10)
+    fun givenTCPPortConfiguration_whenUnixFlags_thenAreNotAddedAsExtras() {
+        val setting = TorConfig.__ControlPort.Builder {
+            unixFlags { GroupWritable = true }
         }
-        assertEquals(10, map["3"])
-        builder!!.HSv3(5)
-        assertEquals(10, map["3"])
+
+        assertEquals("auto", setting.argument)
+        assertEquals(0, setting.optionals.size)
     }
 }
