+++ library/runtime-api/src/commonTest/kotlin/io/matthewnelson/kmp/tor/runtime/api/SocksPortUnitTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.runtime.api.config
+package io.matthewnelson.kmp.tor.runtime.api
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
