+++ library/runtime-api/src/commonTest/kotlin/io/matthewnelson/kmp/tor/runtime/api/builder/UnixSocketBuilderBaseTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.runtime.api.config.builders
+package io.matthewnelson.kmp.tor.runtime.api.builder
 
 import io.matthewnelson.kmp.file.SysPathSep
 import io.matthewnelson.kmp.file.path
